scripts/GSheet/requirements.txt
azure-common==1.1.28
azure-core==1.26.3
azure-identity==1.12.0
azure-keyvault-secrets==4.6.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
google-auth==2.16.0
google-auth-oauthlib==0.8.0
gspread==5.7.2
httplib2==0.21.0
idna==3.4
isodate==0.6.1
msal==1.21.0
msal-extensions==1.0.0
msrest==0.7.1
oauth2client==4.1.3
oauthlib==3.2.2
portalocker==2.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.14
